+++ src/hooks/reactQuery/login/useLoginMutation.ts
@@ -1,5 +1,7 @@
 import { useMutation } from "@tanstack/react-query";
 import { login } from "@/utils/sign";
+import { useAppDispatch } from "@/redux/store";
+import { update } from "@/redux/userSlice";
 
 interface LoginData {
   email: string;
@@ -12,5 +14,14 @@ const fetchLogin = async (data: LoginData) => {
 };
 
 export const useLoginMutation = () => {
-  return useMutation(fetchLogin);
+  const dispatch = useAppDispatch();
+  const {mutate, isLoading} = useMutation(fetchLogin,{
+    onSuccess: uid => {
+      dispatch(update(uid));
+    },
+    onError: () => {
+      alert("로그인 실패했습니다. 다시 로그인 해주세요");
+    },
+  });
+  return {mutate, isLoading}
 };
